+++ PdfSharpDslCore/Parser/VariablesDictionary.cs
@@ -68,6 +68,7 @@ namespace PdfSharpDslCore.Parser
                 case "PAGEHEIGHT":
                 case "PAGEWIDTH":
                 case "PAGECOUNT":
+                case "CURSORY":
                     value = _systemVariablesGet(key);
                     return true;
                 default: